+++ CHANGELOG.md
@@ -17,6 +17,7 @@ _2016-06-03_
 
 * Added support for time based guards. See `AbstractWorkflowTransition::timeHasPassed()`
 * Removed AbstractWorkflowContext::getLastStateChangedAt() and $lastStateChangedAt, you need to implement it in your subclass
+* Removed AbstractWorkflowContext::getStateHistory() - not needed... yet
 
 <a name="version_0_2_0"></a>